% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsloopmix.R
\name{metricsloopmix}
\alias{metricsloopmix}
\title{metricsloopmix}
\usage{
metricsloopmix(d, var, MO, PF = TRUE, s)
}
\arguments{
\item{d}{a data frame.}

\item{var}{a vector of varibales in the data set.}

\item{MO}{is a data frame where:
the first column define the model type, such as "lm" or "lmer"
the second column define the model as "x + y", or "x + y + (1|random)", or "x + y, family = Gamma". Varibales have to be present in the dataset d.
the third column define the predict function.
the fourth column define the type of prediction.}

\item{PF}{if TRUE use the prediction function from the data frame (3rd column of MO object), otherwise it uses "predict".}

\item{s}{significant digits
empty:  return all values.
number 1 to 6: will return the desired significant digits.
"std1": will return number as character in a predefined format.}
}
\value{
a data frame with different parameters to evaluate the model
}
\description{
metricsloopmix
}
