% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opmetrics.R
\name{opmetrics}
\alias{opmetrics}
\title{opmetrics}
\usage{
opmetrics(o, p, s, plot = FALSE, residuals = FALSE)
}
\arguments{
\item{o}{the vector with the observed values}

\item{p}{the vector with the predicted values}

\item{s}{significant digits
empty:  return all values.
number 1 to 6: will return the desired significant digits.
"std1": will return number as character in a predefined format.}

\item{plot}{plot observed against predicted value in ggplot.}

\item{residuals}{if TRUE together with plot, it plots the residuals on the y axis against the predicted values in ggplot.}
}
\value{
a data frame with different parameters to evaluate the predicted
}
\description{
opmetrics
}
