% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSEvectors.R
\name{RMSEvectors}
\alias{RMSEvectors}
\title{RMSEvectors}
\usage{
RMSEvectors(o, p, s)
}
\arguments{
\item{o}{the vector with the observed values}

\item{p}{the vector with the predicted values}

\item{s}{significant digits
empty:  return all values.
number 1 to 6: will return the desired significant digits.
"std1": will return number as character in a predefined format.}
}
\value{
a data frame with different parameters to evaluate the predicted
}
\description{
RMSEvectors
}
