% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfig.R
\name{sigfig}
\alias{sigfig}
\title{signfig}
\usage{
sigfig(vec, digits)
}
\arguments{
\item{vec}{a numeric vector}

\item{digits}{number of significant digits requested}
}
\value{
value(s) as character with the selected number of signiciant digits
}
\description{
signfig
}
